requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.1
Django==1.11.1
djangorestframework==3.7.7
gunicorn==19.6.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
psycopg2==2.6.2
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.2
